(refactor) 게임 참가자 리팩토링
