+++ src/main/java/ladder/domain/player/Players.java
@@ -1,18 +1,23 @@
-package ladder.domain;
+package ladder.domain.player;
 
 import ladder.util.StringUtil;
 
 import java.util.Collections;
 import java.util.List;
+import java.util.concurrent.atomic.AtomicInteger;
 import java.util.stream.Collectors;
 
 public class Players {
     private final List<Player> players;
 
     private Players(String names) {
+        AtomicInteger position = new AtomicInteger(0);
         this.players = StringUtil.convertList(names)
                 .stream()
-                .map(name -> Player.of(name))
+                .map(name -> Player.builder()
+                        .name(name)
+                        .position(position.getAndIncrement())
+                        .build())
                 .collect(Collectors.toList());
     }
 
@@ -22,10 +27,6 @@ public class Players {
 
     public List<Player> getPlayers() {
         return Collections.unmodifiableList(players);
-    }
-
-    public Player getPlayer(int index) {
-        return players.get(index);
     }
 
     public int getCountOfPlayers() {
