+++ src/test/java/ladder/domain/player/PlayerTest.java
@@ -1,5 +1,6 @@
-package ladder.domain;
+package ladder.domain.player;
 
+import ladder.domain.player.Player;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
@@ -12,9 +13,13 @@ class PlayerTest {
     public void name_length_over_exception_test() {
         // given
         String name = "abcdef";
+        int position = 2;
 
         // when & then
-        assertThatThrownBy(() -> Player.of(name))
+        assertThatThrownBy(() -> Player.builder()
+                .name(name)
+                .position(position)
+                .build())
                 .isInstanceOf(IllegalArgumentException.class)
                 .hasMessage("이름은 최대 5글자까지 가능합니다." + name);
     }
