+++ src/test/java/ladder/domain/player/PlayersTest.java
@@ -1,5 +1,6 @@
-package ladder.domain;
+package ladder.domain.player;
 
+import ladder.domain.player.Players;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
